Calculates overview data in python and sends it via JSON object.
Once the page is loaded, an ajax request is made to
/budget/get_overview_data which returns a JSON object.
From there, JavaScript is used to build the table. So
far, this method is significantly faster than the previous
version. Still need to update previous methods of inserting
data into the form and then test for issues.

Some print statements were removed and some form id values were
changed to make it clear what they would be in the DOM for moving
text from the table row to the form.